requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
bcc==0.18.0
behave==1.2.6
blinker==1.4
boto3==1.26.54
botocore==1.29.54
Brlapi==0.8.3
Brotli==1.0.9
cachelib==0.10.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
croniter==1.3.8
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
exceptiongroup==1.1.0
greenlet==1.1.2
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
howdoi @ git+https://github.com/gleitz/howdoi.git@58da8af6913e75a66f85931c137d194eeeca5a7d
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
jmespath==1.0.1
keep==2.10.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.1.0
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
mutagen==1.45.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
outcome==1.2.0
packaging==23.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
pop-transition==1.1.2
protobuf==3.12.4
psutil==5.9.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydbus==0.6.0
PyGithub==1.57
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyquery==2.0.0
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.2.1
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
repolib==2.2.0
repoman==1.4.0
requests==2.25.1
rich==13.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
schedule==1.1.0
SecretStorage==3.3.1
selenium==4.8.0
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==234
terminaltables==3.1.10
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
webdriver-manager==3.8.5
websockets==9.1
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
